Move html parse to util
